requirements.txt
Django==3.2.5
djangorestframework==3.13.1
pytz==2021.3
requests==2.27.1
